restructure the menu components
