+++ src/components/Menu/index.jsx
@@ -1,81 +1,7 @@
-import { MenuItem, MobileThemeSwitcher } from "../MenuItem";
-import styles from "./Menu.module.css";
-import { MenuToggleButton, ThemeSwitcherButton } from "../Button";
-
-const MenuDesktop = ({ menuItems, darkMode, switchTheme, scrollToRef }) => (
-  <div className={styles.menuDesktop}>
-    {menuItems.map((item) => (
-      <MenuItem
-        key={item.title}
-        type="desktop"
-        title={item.title}
-        reference={item.reference}
-        scrollToRef={scrollToRef}
-      />
-    ))}
-    <ThemeSwitcherButton darkMode={darkMode} switchTheme={switchTheme} />
-  </div>
-);
-
-const MenuMobile = ({
-  menuItems,
-  isOpen,
-  toggleMenu,
-  scrollToRef,
-  darkMode,
-  switchTheme,
-}) => {
-  const content = menuItems.map((item) => (
-    <MenuItem
-      key={item.title}
-      type="mobile"
-      title={item.title}
-      icon={item.icon}
-      reference={item.reference}
-      scrollToRef={scrollToRef}
-    />
-  ));
-
-  return (
-    <div className={`${styles.menuMobile} ${isOpen && styles.open}`}>
-      <div className={styles.wrapper}>
-        <MenuToggleButton type="close" toggleMenu={toggleMenu} />
-        <div className={styles.menuMobileContent}>
-          {content}
-          <MobileThemeSwitcher darkMode={darkMode} switchTheme={switchTheme} />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const DeviceMenu = ({
-  menuItems,
-  darkMode,
-  switchTheme,
-  isOpen = false,
-  viewportWidth = 769,
-  toggleMenu,
-  scrollToRef,
-}) => {
-  return viewportWidth > 768 ? (
-    <MenuDesktop
-      menuItems={menuItems}
-      darkMode={darkMode}
-      switchTheme={switchTheme}
-      scrollToRef={scrollToRef}
-    />
-  ) : (
-    <MenuMobile
-      menuItems={menuItems}
-      isOpen={isOpen}
-      toggleMenu={toggleMenu}
-      scrollToRef={scrollToRef}
-      darkMode={darkMode}
-      switchTheme={switchTheme}
-    />
-  );
-};
+import { MenuToggleButton } from "../Button";
+import MenuDesktop from "./MenuDesktop";
+import MenuMobile from "./MenuMobile";
+import DeviceMenu from "./DeviceMenu";
 
 const Menu = ({
   menuItems,
